requirements.txt
annotated-types==0.7.0
anyio==4.6.2
click==8.1.7
colorama==0.4.6
fastapi==0.115.0
fastapi-cli==0.0.5
h11==0.14.0
idna==3.10
psycopg2==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.12
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.37.2
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.31.0
